requirements.txt
numpy>=1.24
sounddevice>=0.4.6
openai>=1.0.0
python-dotenv>=1.0.0
customtkinter>=5.2.0
torch>=2.0.0
whisper
ffmpeg-python>=0.2.0
